DeepGuard/requirements.txt
streamlit
tensorflow>=2.12
keras>=3.0
opencv-python
pillow
matplotlib 